Changed vobulary filtering to work with the spec change.
Vocabulary field might be multiple strings concatinated together, and we need to be able to filter not only by field, but by its substring.

When we filter we want to be able to ignore character case.

Hence 2 extra tests added to cover for this, and the implementation was made.

The culture info had to be surpressed to lower priority as judding from some research, the thing warning was suggesting is out of date with method's actual implementation.

Apparently 'Contains' check using 'Ordinal' by default, so no need to include a 4 word namespace enum.