Added "Get Taxonomy by Id" Controller method base.
Added a test helper for Random value generation. Apparently, when it comes generating single values, AutoFixture is crap.

One worry I have is: since "id" is a mandatory field, if the user makes a request with ".../taxonomies/%20", it typically gets mapped to being null or empty, however in the case of int, that will likely result in "0". Is this what we want?