Addresses API Gateway tests and changes forced by them.
Decided to split the commit of implementation and tests as this is getting too big - as in too many files involved.

The idea here is that for this endpoint, the postcode field is optional, and hence if the part related to it should fail, it shouldn't stop the endpoint from returning a service without distance calculations.

Did few more Randomm class changes to cover for the Addresses API 500 response. Also created exceptions that the gateway should use to notify the usecase that something went wrong. Since the gateway is only expected to return the coordinates, it can't return any messages - hence exceptions.

All exceptions had been made sure to populate the Message field so that the extracting failure information would be as simple as possible.

I better implementation of this is probably possible, where the gateway could instead return some object that has coordinates and Message field, and most of the exceptions should come from the Context instead, however time is pressing and that will have to be done in refactoring.