Added default service sorting to Get Multiple Services endpoint.
If the postcode is provided, services have to be sorted by ascendingly by distance from postcode to service. Presumably sorted by the lowest distance from service locations collection.

If the postcode is not provided, then services have to be sorted ascendingly by service name.

Increased the number of service locations in EntityHelpers so that tests would have more to work with. It's too easy to have 3 services sorted by random chance.