Added Exception Thrower test helper class
Allows for exception triangulation as well as distinguishing between regular exceptions and aggregate exceptions.

Changed the warning level of the CA1031 notification from warning (which will prevent the application from compiling), to suggestion. The reason why it's okay to do this is because the warning says that it's better to leave the performance and similar related errors uncaught because it will aid in debugging. In our case, all we're doing is simply reformating how the exception message will look like, which means we're these exceptions do not get lost in the process of error handling, which means we're not in conflict with the reason CA1031 was introduced.