A hack of ignoring services who's organization was deleted.
At this point in time the Portal API's DELETE an organization endpoint removes an organization from the database, and sets all the organization id's of related services to null. This results in the database containing such orphan services with no related organization.

It was discovered that when the Public (this) API does a query to get all the services, it crashes upon trying to retrieve a collection of services that contains a service related to an organization, which doesn't exist.

The hack is that rather than addressing the DELETE endpoint behaviour, as the bussiness logic is unclear (waiting for a response on that behavior), we instead 'fix' the Public API so it would retrieve only active services. So it's the case of treating the symptom rather than the cause.

If the organization providing the services is gone, I'm not sure why we're keeping the service. But that's what we're waiting on to find out.