Made the search check for partial matches.
This works fine for the case when there are no full matches between what's in the DB and what user inputed. However this still has an issue of polluting the results, when the matches are present - due to the sorting functionallity the less relevant (partial match) results will be sorted into the more relevant results found by the full match.

In addition to that, the result pollution can grow pretty high if user input something like "a park" - in this case, the search would return all the DB entries that contain the character 'a'.