Changed the Search Services Gateway method Contract and TESTS
This change stems from the need of having more relevant user input matches be separated from less relevant matches, where the input was split into pieces.

Since the gateway tests had to be changed anyway, I decided to change them in a way where they would test the final intended functionality rather than do a temporary change to get them work with the new contract.

Many other tests and pieces of implementation are missing in this commit. I'm splitting this commit out to have more managable chunks of changes. An alternative would be having close to double the changes seen here, which would be hard to read an describe under a single commit.