Added random SearchServiceGatewayResult generator method
Due to the change in that gateway's contract, many tests need be changed so they would compile. Having this block of setup in each of those tests would be an issue both in terms of DRY and readability.

Generates random number of services for the sake of triangulation. Generates no less than 5 services to ensure relatively low chance of the services being generated in a sorted order, which would throw off the sorting tests.