Changed the SearchService endpoint ToResponse factory converter method.
Due to the extra requirements of how a new search has to work, the gateway has to return 2 lists - higher relevance one and lower relevance one. They have to be separate due to that gateway unrelated postprocessing reasons.

Previous factory method was built with the assumption that the gateway in question will only return 1 list, which is no longer true. Hence the approach of mapping the data into the final response object has to be changed.

The previous way was hacky already, but there's no way to resolve that due to intermingled infrastructure entity and domain object trees and incomplete domain objects. Part of this solution (where it accepts list of service response rather than list of service domain) still suffers from that problem, however solving of that has to be postponed as it would take forever due to having to resolve root issue (it requires a separate PR).