Changes to Search Services endpoint's usecase - sorting, mapping.
Changed how sorting gets done on the usecase. Inb4 it would simply sort the only list it would get. However now that there are 2 lists, they have to be sorted individually and joined together - which in effect makes it so the full list doesn't follow ascending pattern.

Also changed how the mapping gets done. Unfortunatelly, due to intermingled Domain and Entity object tree's, there's little benefit to go and edit the ServiceLocation domain object so it would contain the distance field. I need that field so that I wouldn't have to manually map each gateway result list to response object form, which contains the distance field.

In addition to that, changed the approach with the metadata object. Rather than generating an object with nothing but nulls within usecase response. I can shorten the lines of code by creating the metadata in the usecase, and passing it into factory.

Finally, I extracted the distances calculation logic into a separate method so that the main method would look cleaner, plus I didn't want to have it repeated twice in the same method. This helper method still needs refactoring, so it could be reused in getting a single service. Since this is unfinished business, I might as well leave this helper class here in the usecase until upcoming refactoring.

Thing to consider - how does sorting by distance work, when among all the services, there is 1 service with no service locations? That's a missing test - a TODO.